Create one reducer for all sessions
