+++ src/redux/user/user.js
@@ -3,7 +3,6 @@ import axios from 'axios';
 const BASE_URL = 'http://localhost:3001/api';
 const REGISTER_URL = 'register';
 const LOGIN_URL = 'login';
-const LOGOUT = 'userLogOut';
 
 export const register = async (name, email, password, passwordConfirmation) => {
   const data = await axios({
@@ -36,9 +35,3 @@ export const initialState = {
   fetchedData: '',
   status: '',
 };
-
-export const userLogOutAction = () => ({
-  type: LOGOUT,
-});
-
-export { LOGOUT };
